% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_calculator.r
\name{ds_calculator}
\alias{ds_calculator}
\title{Calculator for d<sub>s</sub> (between subjects effect size)}
\usage{
ds_calculator(group1, group2)
}
\arguments{
\item{x, y}{vectors of the DV for each group}
}
\description{
Calculator for d<sub>s</sub> (between subjects effect size)
}
\examples{
ds_calculator()
}
