% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dz_calculator.r
\name{dz_calculator}
\alias{dz_calculator}
\title{Calculator for d<sub>z</sub> (within subjects effect size)}
\usage{
dz_calculator(x, y)
}
\arguments{
\item{x, y}{vectors of the DV for each level of the IV}
}
\description{
Calculator for d<sub>z</sub> (within subjects effect size)
}
\examples{
dz_calculator()
}
