% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixels_to_dva.R
\name{pixels_to_dva}
\alias{pixels_to_dva}
\title{Degrees of Visual Angle Calculator}
\usage{
pixels_to_dva(pixels, screen_res = c(1920, 1080), view_distance = 60,
  monitor_dims = c(50.92, 28.64))
}
\arguments{
\item{pixels}{size of the object in pixels}

\item{screen_res}{screen resolution, e.g. 1920 x 1080 = c(1920, 1080)}

\item{view_distance}{viewing distance in cm}

\item{monitor_dims}{width and height of the monitor in cm, e.g. c(50.92, 28.64)}
}
\description{
Calculates degrees of visual angle from pixels, screen resolution, viewing distance, and monitor size
}
