% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySEwithin2.r
\name{summarySE2}
\alias{summarySE2}
\title{Calculate standard error}
\usage{
summarySE2(data = NULL, measurevar, groupvars = NULL, na.rm = TRUE,
  conf.interval = 0.95)
}
\description{
Calculate standard error
}
