% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySEwithin2.r
\name{summarySEwithin2}
\alias{summarySEwithin2}
\title{Calculate within-subjects standard error}
\usage{
summarySEwithin2(data = NULL, measurevar, betweenvars = NULL,
  withinvars = NULL, idvar = NULL, na.rm = TRUE,
  conf.interval = 0.95)
}
\description{
Calculate within-subjects standard error
}
