% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dp.R
\name{theme_dp}
\alias{theme_dp}
\title{Personal Light Theme}
\usage{
theme_dp(base_family = "Roboto Condensed", base_size = 12,
  plot_title_family = base_family, plot_title_size = 18,
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = base_family, subtitle_size = 12,
  subtitle_face = "plain", subtitle_margin = 15,
  strip_text_family = base_family, strip_text_size = 12,
  strip_text_face = "plain", caption_family = base_family,
  caption_size = 9, caption_face = "italic", caption_margin = 10,
  axis_text_size = base_size, axis_title_family = subtitle_family,
  axis_title_size = base_size, axis_title_face = "plain",
  axis_title_just = "cc", plot_margin = margin(30, 30, 30, 30),
  grid_col = "#cccccc", grid = FALSE, axis_col = "#000000",
  axis = TRUE, ticks = TRUE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot title family, face, size and margi}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin()])}

\item{grid_col, axis_col}{grid & axis colors; both default to `#cccccc`}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}
}
\description{
Personal [ggplot2] theme for manuscript figures, based heavily on the package: hrbrthemes (https://github.com/hrbrmstr/hrbrthemes)
}
\examples{
theme_dp()
}
