% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmac_pal.R
\name{vmac_pal}
\alias{vmac_pal}
\title{Qualitative colour palette for standard VMAC colours}
\usage{
vmac_pal(high = TRUE, low = TRUE, absent = TRUE, other = FALSE)
}
\arguments{
\item{high, low, absent, other}{TRUE/FALSE to pull colour for each trial type}
}
\description{
Qualitative colour palette for standard VMAC colours
}
\examples{
vmac_pal()
}
